requirements.txt
python-telegram-bot==13.14
openai==0.26.4
requests==2.28.1
Pillow==9.5.0